feat(商家端): 重构营销活动页面，完善优惠券管理功能
## 商家端前端

### 营销活动页面重构 (Campaign.vue)
- 新增统计概览卡片：活跃优惠券数、领取量、核销率、带来营收
- 新增标签页筛选：全部/生效中/已过期/待生效，支持按状态切换
- 新增优惠券列表卡片式布局，展示类型标签、折扣金额、使用门槛、有效期
- 新增分页功能：支持上一页/下一页/跳转指定页
- 新增操作功能：编辑、启用/停用切换、删除优惠券
- 新增辅助函数：getCouponTypeText/getDiscountDisplay/getThresholdDisplay/getValidityDisplay/getStatusText等

### API模块更新 (promo.ts)
- 重构API路径：/api/admin/promo/coupon/* → /api/admin/coupon/*
- 重构API路径：/api/admin/promo/group-buy/* → /api/admin/group-buy/*
- 列表接口路径从 page 改为 list

### 类型定义更新 (promo.ts)
- 字段重命名：threshold → minAmount，totalCount → total，claimedCount → receivedCount
- 新增字段：createdAt、updatedAt

### 路由配置 (router/index.ts)
- 新增优惠券创建页面路由：/campaign/coupon/create → CouponCreate.vue

## 用户端前端

### 分类页优化 (category.vue)
- 错误日志添加中文描述，便于调试定位

## 后端

### 优惠券实体 (Coupon.java)
- scopeProductIds 类型从 List<Long> 改为 List<Integer>，与数据库整型字段保持一致

+++ xianguoji-server/src/main/java/com/xianguoji/server/module/promo/entity/Coupon.java
@@ -31,7 +31,7 @@ public class Coupon {
     private Integer validDays;
     private Integer scope;
     @TableField(typeHandler = JacksonTypeHandler.class)
-    private List<Long> scopeProductIds;
+    private List<Integer> scopeProductIds;
     private Integer status;
     private LocalDateTime createdAt;
     private LocalDateTime updatedAt;
